got base mode

+++ src/components/FeedBackItem/FeedbackItem.js
@@ -9,11 +9,15 @@ class FeedbackItem extends Component {
 
     deleteFeedback = () => {
         console.log(this.props.feedback.id);
+        if (window.confirm('Are you sure you want to delete thie feedback forever (forever...forever...)?')) {
             axios.delete(`/api/feedback/${this.props.feedback.id}`, this.props.feedback.id).then((response) => {
                 this.props.getFeedbackFromServer();
             }).catch((error) => {
                 alert('problemssssss');
             });
+        } else {
+            console.log('do NOT delete');
+        }
     }
 
     render() {
